requirements.txt
asgiref==3.11.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Django==6.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
sqlparse==0.5.4
tzdata==2025.2
Werkzeug==3.1.4
pymongo
python-dotenv
Flask-CORS
google-generativeai
